chore: 라이브 코딩 전에 코드 정리 (#36)
* simplify name

* remove unneccesary code

* Move sampleComponent into another file
