+++ src/index.tsx
@@ -1,23 +1,3 @@
-import * as React from 'react';
-
-interface IProps {
-  text: string;
-}
-
-const SampleComponent = (props: IProps) => {
-  const {text} = props;
-  const [value] = React.useState('Ready to use React Hook!');
-  return (
-    <>
-      <div style={{backgroundColor: 'black', color: 'white'}}>
-        {text}
-        {value}
-      </div>
-    </>
-  );
-};
-
-export default SampleComponent;
 export * from './components/AnalyticsProvider';
 export * from './utils/googleAnalytics';
 export * from './contexts/useAnalyticsContext';
