+++ src/utils/googleAnalytics/event.ts
@@ -1,6 +1,6 @@
-import { gtag } from "./initialize";
+import {gtag} from './initialize';
 
-export function logEvent(name: string, params?: { [key: string]: string }) {
+export function event(name: string, params?: {[key: string]: string}) {
   console.info(`✅GA: event ${name}`, params);
-  gtag("event", name, params);
+  gtag('event', name, params);
 }
